Move errors into their respective crates
