+++ linkerd/app/gateway/src/tests.rs
@@ -1,12 +1,12 @@
 use super::*;
 use linkerd_app_core::{
     dns,
-    errors::{GatewayDomainInvalid, GatewayIdentityRequired, GatewayLoop},
     identity as id, profiles,
     proxy::http,
     svc::NewService,
     tls, Error, NameAddr, NameMatch,
 };
+use linkerd_app_inbound::{GatewayDomainInvalid, GatewayIdentityRequired, GatewayLoop};
 use linkerd_app_test as support;
 use std::str::FromStr;
 use tower::util::ServiceExt;
