+++ linkerd/app/outbound/src/http/mod.rs
@@ -7,6 +7,7 @@ mod server;
 
 use crate::tcp;
 pub use linkerd_app_core::proxy::http::*;
+pub(crate) use self::require_id_header::IdentityRequired;
 use linkerd_app_core::{
     dst,
     profiles::{self, LogicalAddr},
